added font config

+++ dotfiles/install.sh
@@ -6,16 +6,17 @@ typeset -A files
 
 THIS_DIR=$HOME/.vim/dotfiles
 
-files=( bashrc        $HOME/.bashrc \
-        clang-format  $HOME/.clang-format \
-        gitconfig     $HOME/.gitconfig \
-        gvimrc        $HOME/.gvimrc \
-        screenrc      $HOME/.screenrc \
-        tmux.conf     $HOME/.tmux.conf \
-        vimrc         $HOME/.vimrc \
-        Xmodmap       $HOME/Xmodmap \
-        Xresources    $HOME/.Xresources \
-        zshrc         $HOME/.zshrc \
+files=( bashrc           $HOME/.bashrc \
+        clang-format     $HOME/.clang-format \
+        gitconfig        $HOME/.gitconfig \
+        gvimrc           $HOME/.gvimrc \
+        screenrc         $HOME/.screenrc \
+        tmux.conf        $HOME/.tmux.conf \
+        vimrc            $HOME/.vimrc \
+        Xmodmap          $HOME/Xmodmap \
+        Xresources       $HOME/.Xresources \
+        zshrc            $HOME/.zshrc \
+        fonts_local.conf $HOME/.config/fontconfig/fonts.conf \
       )
 
 #compton.conf
@@ -35,11 +36,16 @@ for f in "${(@k)files}"; do
         if [[ -f $TARGET && ! -h $TARGET ]]; then
             cp $TARGET $TARGET.old
             echo "backed up $TARGET to $TARGET.old"
+            rm -f $TARGET
         fi
-        rm -f $TARGET
     fi
 
-    ln -s $SOURCE $TARGET
+
+    if [[ ! -f $TARGET ]]; then
+        ln -s $SOURCE $TARGET
+    else
+        echo "skipping $f"
+    fi
 done
 
 echo "DONE!"
